Tube Unit: Work toward smooth root frequency adjustment.
It's not working yet, so turned off, but reworking the
Tube Unit engine to be able to smoothly adjust the
root frequency, instead of being limited to an integer
number of samples in the tube length.

+++ src/sapphire_engine.hpp
@@ -569,7 +569,7 @@ namespace Sapphire
                 // For now, use modified sinc() with a simple cosine window.
                 // FIXFIXFIX: Consider replacing with a Kaiser window:
                 // https://ccrma.stanford.edu/~jos/pasp/Theory_Practice.html
-                double taper = sinc * cos(angle / (steps + 1.0));
+                float taper = sinc * cos(angle / (steps + 1.0));
                 sum += buffer[n+s] * taper;
             }
